@@ -5,7 +5,7 @@ author: David Benitez <dbr.bcn@gmail.com>
 homepage: https://github.com/dbenitez-bcn/solid_bottom_sheet
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
